analyzer: Move NuGet service index related classes to their own file
The service index model classes do not belong to the package model
classes as they describe nothing about a package, but where to get
package metadata from.
