+++ analyzer/src/main/kotlin/managers/utils/NuGetServiceIndex.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2022 The ORT Project Authors (see <https://github.com/oss-review-toolkit/ort/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.ossreviewtoolkit.analyzer.managers.utils
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+import com.fasterxml.jackson.annotation.JsonProperty
+
+// See https://docs.microsoft.com/en-us/nuget/api/service-index.
+@JsonIgnoreProperties(ignoreUnknown = true)
+data class ServiceIndex(
+    val version: String,
+    val resources: List<ServiceResource>
+)
+
+// See https://docs.microsoft.com/en-us/nuget/api/service-index#resource.
+@JsonIgnoreProperties(ignoreUnknown = true)
+data class ServiceResource(
+    @JsonProperty("@id")
+    val id: String,
+    @JsonProperty("@type")
+    val type: String
+)
